# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ticket API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/departments/GetAllActiveDepartments/": {
            "get": {
                "description": "Returns a list of all active departments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Get all active departments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DepartmentDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/files/GetDownloadLinkTicketFile/{objectName}/": {
            "post": {
                "description": "Generates a temporary presigned URL for the given file and redirects the client to that URL to start the download.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TicketFile"
                ],
                "summary": "Download a ticket file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File object name (UUID + extension)",
                        "name": "objectName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ticket ID for file location",
                        "name": "ticketId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IDRequestString"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Give MinIO download URL"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/files/UploadTicketFile/": {
            "post": {
                "description": "Upload a file to the temporary storage for a ticket. File must be multipart/form-data with field name `file`.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TicketFile"
                ],
                "summary": "Upload a ticket file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Ticket file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns uploaded file ID",
                        "schema": {
                            "$ref": "#/definitions/dto.IDResponseString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "415": {
                        "description": "Unsupported file extension",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/otp/send": {
            "post": {
                "description": "Generates a verification code, caches in Redis, logs to SMS warehouse, and sends SMS via provider asynchronously",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OTP"
                ],
                "summary": "Send OTP verification code",
                "parameters": [
                    {
                        "description": "Phone number",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendOTPDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification code sent",
                        "schema": {
                            "$ref": "#/definitions/dto.SendOTPResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "description": "Validates provided OTP code against cached code in Redis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OTP"
                ],
                "summary": "Verify OTP code",
                "parameters": [
                    {
                        "description": "Phone number and OTP code",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyOTPDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP is valid",
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyOTPResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errx.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DepartmentDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.IDRequestString": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.IDResponseString": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.SendOTPDTO": {
            "type": "object",
            "required": [
                "phoneNumber"
            ],
            "properties": {
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "dto.SendOTPResponseDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.VerifyOTPDTO": {
            "type": "object",
            "required": [
                "code",
                "phoneNumber"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "dto.VerifyOTPResponseDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "errx.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errx.Error"
                }
            }
        },
        "errx.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errx.ErrorCode"
                },
                "debug": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "stack": {
                    "type": "string"
                }
            }
        },
        "errx.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29
            ],
            "x-enum-varnames": [
                "ErrInternalServerError",
                "ErrTicketNotFound",
                "ErrUnauthorized",
                "ErrInvalidInput",
                "ErrDuplicate",
                "ErrBadRequest",
                "ErrUserNotFound",
                "ErrTicketTypeNotFound",
                "ErrDepartmentNotFound",
                "ErrUserDuplicate",
                "ErrInvalidCredentials",
                "ErrWeakJWTSecret",
                "ErrIncorrectCaptcha",
                "ErrExpiredCaptcha",
                "ErrTicketStatusNotFound",
                "ErrTooManyRequest",
                "ErrApiKeyNotFound",
                "ErrServiceUnavailable",
                "ErrFileNotFound",
                "ErrLinkExpired",
                "ErrUnsupportedFileExtension",
                "ErrMaxFileSizeExceeded",
                "ErrMaxTicketFilesExceeded",
                "ErrRequestBodyTooLarge",
                "ErrForbidden",
                "ErrNotFound",
                "ErrOTPExpired",
                "ErrOTPInvalid",
                "ErrOTPSendFailed",
                "ErrOTPMaxAttemptsExceeded"
            ]
        }
    }
}